removing slow code on profile page
Removing code which is slowing down the profile page. This code went and checked if a staff member's photo exists on www2 and if it doesn't exist displays a generic image. However, this process is slowing down the page load considerable. Therefor, I have removed it and instead added a thumbnail for all staff - generic ones for those without images.