align label/url for related resources (#1640)
align label/url for related resources

* use SymphonyService to populate electronic access
* add feature tests for electronic access links and record vcr cassettes for ruby http requests 
* add three new marc records to sample set for UAL access, finding aid 
and on-campus access conditions
* fix helper for electronic resources

Realized that there are three categories of electronic resources we 
consume. Primary (40/41 indicators) which can be 1) UAL access or Free 
or 2) Not UAL access, and 3) Supplimental (42 indicators) for things 
like finding aids. 1 and 3 appear at the top of the page while 2 appears at the bottom.