use safe operator on empty location_tesim
Searches that contain results with empty `location_tesim` were failing
since September 11, 2019 at ~9:30am.  There was a re-index shortly before
that date but not a direct corelation. The cause at this time and
any other implications is unknown.

location_tesim is used to show the item's copies owned by. It also is a
facet and used to detect if the UAL shield should be shown.