prepare for incoming library short codes
Working towards automating changes to the human readable strings for
the 'Where is this?' table. Locations belong to a Library.  I'm not sure
of the origin of the current short codes but it looks like it's been
a challenge to deal with https://github.com/ualbertalib/NEOSDiscovery/commit/4da3e71979285f9548526d7a2dd1465b690d1cb1#diff-e678c8aa2cbba5c5a1251bc942a2182a

There are some values that we have in our tables currently (url, neos_url and proxy)
which don't have an analog in the Symphony configuration. If the Symphony
configuration is driving the nightly update then we'll need a way to map
between these two names for the same thing.

I've put together this table to illustrate what I'm thinking makes sense.
I've joined the libraries based on a common proxy url between the export
provided and the existing configuration.
|LIBG|proxy|existing|
|--|--|--|
|ALL_LIBS
|UALB_CAT
|VANGUARD_|VANGUARD|vanguard|
|AGL_|AGINTERNET http://login.ezproxy.library.ualberta.ca/login?url=|universityofalberta|
|AHS_|AHSGLENRSE http://ahs.idm.oclc.org/login?url=|abhealth|
|AITF_||innovates|
|CONCORDIA_|CONCORDIA http://aec.talonline.ca/login?url=|concordia|
|COVENANT_|COVENANTGN http://ahs.idm.oclc.org/login?url=,|covenant|
|UALBERTA2|
|BUR_|CUC http://ezproxy.achcu.talonline.ca/login?url=|burmanuniversity|
|GPRC_|GPRC_GP https://ezproxy.agpc.talonline.ca/login?url=|gp|
|MACEWAN_|GR_MACEWAN http://ezproxy.macewan.ca/login?url=|macewanuniversity|
|NEOS_FREE_||neos|
|KEYANO_|KEYANO http://ezproxy.afmk.talonline.ca/login?url=|keyano|
|KINGS_|KINGS http://aekc.talonline.ca/login?url=|kings|
|LAKELAND_|LAKELND_VR http://ezproxy.avc.talonline.ca/login?url=|lakeland|
|NEWMAN_|NEWMAN http://ezproxy.newman.edu:81/login|newman|
|OLDS_|OLDS http://ezproxy.aoac.talonline.ca/login?url=|olds|
|RDC|RED_DEER_C http://ardc.talonline.ca/login?url=|reddeer|
|TAYLOR_|
|UINTERNET_|
|NORQUEST_|NORQ_MAIN http://auth01.norquest.ca/login?url=|norquest|
|NLC_|NLC_SL http://ezproxy.asav.talonline.ca/login?url=|nlc|
|ALL_LIBS2|
|UARCRF_|
|UALBERTA|
||ARC_MW|
|||gov https://ezproxy.aee.talonline.ca/login?url=|
|||free|

I'm keeping the old ones because they might be necessary in NEOSDiscovery.