denormalize library locations
Working towards automating changes to the human readable strings for
the 'Where is this?' table.

There is one value in our table currently (url) which doesn't have an
analog in the Symphony configuration. If the Symphony configuration is
driving the nightly update then we'll need a way to map between these
two names for the same thing.

My observation is that the normalization that's being done isn't adding
any value (needs conversion on both sides), so we might as well just
store the value that Symphony is using.