create backup tables for library and location
Working towards automating the nightly populating of these tables.  I'd
like to have a backup for a couple reasons:
1. if something goes sideways during the migration we can just restore the old tables
2. there is some data in the table which cannot be automated (urls, relationships between location and library)

For reason 2 I'm also adding ActiveRecord classes to help navigate the backup data.
These need to exist before the symphony_nightly generator is run even if they're
empty to begin with.

- magic of Rails means that the table names do NOT need to be specified
- belongs_to is smart about foreign key names
- added tests to have confidence that I was invoking the magic correctly