change to incoming data makes somethings unnecessary
The Data4DiscoveryManual.txt file will contain the relationships between library locations and their NEOS library parent so we don't need to access the BackupLocations and BackupLibraries.  The library short codes are also not changing so the migration to do so is no longer necessary.

Also some minor tweaks to the erb template so that formatting is nicer.