add solr_marc_prep generator consumed by the ingest rake task
Since release 3.0.113, we have moved the management of location codes,
circulation rules, statuses, and item types into DB, instead of static
yml file in Discovery. This will avoid a new release every time a change
is introduced in Symphony, and will allow for more efficient/accurate
data used in Discovery.

Now we need to use this same information when indexing/ingesting the
MARC data into Solr. When `rake ingest[collection]` is invoked, the
generator will be called which will use the current database values in
a template to replace two files that map to human readable strings for
Institution and Library in the application's facets.