discourage holds placed on items with HathiTrust access
Our initial understanding of HathiTrust's ETAS programme was that when we restored access to the print collection we would have to remove the HathiTrust ETAS links (i.e. ETAS is only allowed when there is no access to the print collection). Clarification from HathiTrust sent out on Friday indicates that access to the print collection can be restored as long as ETAS items are not accessible.

 we're going to suppress the hold button in Discovery whenever the record is in the overlap table, without regard for whether or not the item is actually still available via the HathiTrust API, and knowing that the overlap report is not currently being refreshed in a timely manner.

To be clear: this means that we 100% expect that there will be corner cases in which the hold button will be unavailable but no access in HathiTrust button will be present, because the item is unavailable through their API. This is not a bug, this is a compromise we are making to do this in a timely manner – users are expected to contact public services to resolve access to those items. Weiwei will communicate this expectation to public services.