added oclc number to symphony indexing
Concordia library uses WorldCat Local, and OCLC does not index our 001 cataloguing key. This means, they can't link directly to an item for hold-placing purposes. They can use the OCLC number, which is indexed in WorldCat Local, but we currently don't index that field. If we index that field, we may be able to come up with a solution for them to build a working place hold link.